feat: Reflect property to attribute only if enabled
BREAKING CHANGE: The default value of the `reflect` flag has changed to
`false`. Properties do not reflect their values to attributes
automatically. The default roles of attributes is to initialize
properties, not necessarily to always contain the same value. If you
need to always synchronize the value of a property and an attribute, set
the flag `reflect` to `true` explocitly in the property declaration.